+++ insanic/middleware.py
@@ -1,11 +1,9 @@
-from sanic.log import log
-
 # @app.middleware('request')
 async def request_middleware(request):
 
     pass
 
 async def response_middleware(request, response):
-
+    print("help")
     pass