+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Changelog for insanic
 0.0.185 (unreleased)
 --------------------
 
-- Nothing changed yet.
+- MAJOR: updated sanic to 0.7.0
+- BREAKING: when running insanic in run.py remove log_config parameter
+- FEATURE: better organization of logging modules
+- FIX: 204 empty body assertion in test_api_endpoint
+- REFACTOR: remove log email handlers and reporters
 
 
 0.0.184 (2018-02-07)